// src/types.ts
export type Field<Value> = {
  name: string;
  $initValue: StoreWritable<Value>;
  $value: StoreWritable<Value>;
  $errors: StoreWritable<ValidationError<Value>[]>;
  $firstError: Store<ValidationError<Value> | null>;
  $errorText: Store<string>;
  $isValid: Store<boolean>;
  $isDirty: Store<boolean>;
  $isTouched: StoreWritable<boolean>;
  $touched: StoreWritable<boolean>;
  $field: Store<FieldData<Value>>;
  onChange: EventCallable<Value>;
  changed: EventCallable<Value>;
  onBlur: EventCallable<void>;
  addError: EventCallable<{ rule: string; errorText?: string }>;
  validate: EventCallable<void>;
  reset: EventCallable<void>;
  set: EventCallable<Value>;
  resetErrors: EventCallable<void>;
  resetValue: EventCallable<void>;
  filter?: StoreWritable<boolean> | Store<boolean> | FilterFunc<Value>;
  "@@unitShape": () => FieldUnitShape<Value>;
};
// ...
export type AnyFormValues = {
  [key: string]: any;
};

export type FormValues<Fields extends AnyFields> = {
  [K in keyof Fields]: Fields[K] extends Field<infer U> ? U : never;
};

export type FormFieldConfigs<Values extends AnyFormValues> = {
  [K in keyof Values]: FieldConfig<Values[K]>;
};

export type FormFields<Values extends AnyFormValues> = {
  [K in keyof Values]: Field<Values[K]>;
};

export type AddErrorPayload = {
  field: string;
  rule: string;
  errorText?: string;
};
// ...
export type FormUnitShape<Values extends AnyFormValues> = {
  isValid: Store<boolean>;
  isDirty: Store<boolean>;
  touched: Store<boolean>;
  submit: EventCallable<void>;
  validate: EventCallable<void>;
  reset: EventCallable<void>;
  addErrors: EventCallable<AddErrorPayload[]>;
  setForm: EventCallable<Partial<Values>>;
  setInitialForm: EventCallable<Partial<Values>>;
  resetTouched: EventCallable<void>;
  resetValues: EventCallable<void>;
  resetErrors: EventCallable<void>;
  formValidated: EventCallable<Values>;
};

export type Form<Values extends AnyFormValues> = {
  fields: FormFields<Values>;
  $values: StoreWritable<Values>;
  $eachValid: StoreWritable<boolean>;
  $isValid: StoreWritable<boolean>;
  $isDirty: StoreWritable<boolean>;
  $touched: StoreWritable<boolean>;
  $meta: StoreWritable<{
    isValid: boolean;
    isDirty: boolean;
    touched: boolean;
  }>;
  submit: EventCallable<void>;
  validate: EventCallable<void>;
  reset: EventCallable<void>;
  addErrors: EventCallable<AddErrorPayload[]>;
  set: EventCallable<Partial<Values>>;
  setForm: EventCallable<Partial<Values>>;
  setInitialForm: EventCallable<Partial<Values>>;
  resetTouched: EventCallable<void>;
  resetValues: EventCallable<void>;
  resetErrors: EventCallable<void>;
  formValidated: EventCallable<Values>;
  "@@unitShape": () => FormUnitShape<Values>;
};